doof doof

+++ migrations/2022-09-26-160118_create_chores/up.sql
@@ -2,7 +2,7 @@ CREATE TABLE chores (
   id SERIAL PRIMARY KEY,
   name VARCHAR NOT NULL,
   description VARCHAR,
-  expected_duration_minutes NUMERIC,
+  expected_duration_minutes REAL,
   household_id SERIAL NOT NULL,
   created_by_family_member_id SERIAL,
   CONSTRAINT fk_chores_household
